laptop commit: jwt authorization added for all calls. TODO: test all

+++ src/main/java/raf/edu/rs/projekat/service/UserService.java
@@ -4,6 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
+import org.springframework.http.HttpStatus;
 import org.springframework.orm.ObjectOptimisticLockingFailureException;
 import org.springframework.scheduling.TaskScheduler;
 import org.springframework.scheduling.annotation.Scheduled;
@@ -13,51 +14,97 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import raf.edu.rs.projekat.controller.ApiException;
+import raf.edu.rs.projekat.model.DecodedToken;
 import raf.edu.rs.projekat.model.User;
+import raf.edu.rs.projekat.repository.MachineRepository;
 import raf.edu.rs.projekat.repository.UserRepository;
 
+import java.io.UnsupportedEncodingException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+import java.util.logging.Logger;
 
 @Service
 public class UserService implements IService<User, Long>, UserDetailsService {
 
     private final PasswordEncoder passwordEncoder;
-
+    private Logger logger = Logger.getLogger(this.getClass().getName());
+    //private final MachineRepository machineRepository;
     private final UserRepository userRepository;
+    private DecodedToken token;
 
     @Autowired
     public UserService(PasswordEncoder passwordEncoder, UserRepository userRepository) {
         this.passwordEncoder = passwordEncoder;
-
+        this.token = null;
         this.userRepository = userRepository;
     }
 
-    public User create(User user) {
-        user.setPassword(this.passwordEncoder.encode(user.getPassword()));
-        return this.userRepository.save(user);
+    public User create(User user, String jwt) throws UnsupportedEncodingException {
+        logger.info("CREATE USER...");
+        token = DecodedToken.getDecoded(jwt);
+        if (token.can_create_users) {
+            user.setPassword(this.passwordEncoder.encode(user.getPassword()));
+            return this.userRepository.save(user);
+        } else {
+            throw new ApiException(HttpStatus.FORBIDDEN, "can_create_users privilege missing", null);
+        }
+
+    }
+
+    public User update(User user, String jwt) throws UnsupportedEncodingException {
+        logger.info("UPDATE USER...");
+        token = DecodedToken.getDecoded(jwt);
+        if (token.can_update_users) {
+            return this.userRepository.save(user);
+        } else {
+            throw new ApiException(HttpStatus.FORBIDDEN, "can_update_users privilege missing", null);
+        }
+
     }
 
     @Override
-    public <S extends User> S save(S var1) {
-        var1.setPassword(this.passwordEncoder.encode(var1.getPassword()));
+    public <S extends User> S save(S var1, String jwt) {
+        //var1.setPassword(this.passwordEncoder.encode(var1.getPassword()));
         return userRepository.save(var1);
     }
 
     @Override
-    public Optional<User> findById(Long var1) {
-        return userRepository.findById(var1);
+    public Optional<User> findById(Long var1, String jwt) throws UnsupportedEncodingException {
+        logger.info("READ USER...");
+        token = DecodedToken.getDecoded(jwt);
+        if (token.can_read_users) {
+            return userRepository.findById(var1);
+        } else {
+            throw new ApiException(HttpStatus.FORBIDDEN, "can_read_users privilege missing", null);
+        }
+
+
     }
 
     @Override
-    public List<User> findAll() {
-        return userRepository.findAll();
+    public List<User> findAll(String jwt) throws UnsupportedEncodingException {
+        logger.info("READ USERS...");
+        token = DecodedToken.getDecoded(jwt);
+        if (token.can_read_users) {
+            return userRepository.findAll();
+        } else {
+            throw new ApiException(HttpStatus.FORBIDDEN, "can_read_users privilege missing", null);
+        }
+
     }
 
     @Override
-    public void deleteById(Long var1) {
-        userRepository.deleteById(var1);
+    public void deleteById(Long var1, String jwt) throws UnsupportedEncodingException {
+        logger.info("DELETE USERS...");
+        token = DecodedToken.getDecoded(jwt);
+        if (token.can_delete_users) {
+            userRepository.deleteById(var1);
+        } else {
+            throw new ApiException(HttpStatus.FORBIDDEN, "can_delete_users privilege missing", null);
+        }
     }
 
     @Override
